Switch timeout value to a long
git-svn-id: https://svn.apache.org/repos/asf/openejb/trunk/openejb3@951777 13f79535-47bb-0310-9956-ffa450edef68

+++ container/openejb-jee/src/main/java/org/apache/openejb/jee/Timeout.java
@@ -27,16 +27,16 @@ import javax.xml.bind.annotation.XmlType;
 public class Timeout {
 
 	@XmlElement(name = "timeout", required = true)
-	private Integer timeout;
+	private long timeout;
 
 	@XmlElement(name = "unit", required = true)
 	private String unit;
 
-	public Integer getTimeout() {
+	public long getTimeout() {
 		return timeout;
 	}
 
-	public void setTimeout(Integer timeout) {
+	public void setTimeout(long timeout) {
 		this.timeout = timeout;
 	}
